refactor(run): keep canonical recording policy within package budget
