+++ pkg/transports/cli/run/invocation_hosted_runtime.go
@@ -10,6 +10,7 @@ import (
 	"sort"
 	"strings"
 
+	"github.com/google/uuid"
 	"github.com/portpowered/infinite-you/pkg/initializer"
 	"github.com/portpowered/infinite-you/pkg/initializer/runtimeapplication"
 	"github.com/portpowered/infinite-you/pkg/platform/runtimeartifact"
@@ -25,6 +26,28 @@ import (
 	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
 	"go.uber.org/zap"
 )
+
+func prepareCanonicalSessionIDForRun(cfg RunConfig) (RunConfig, error) {
+	if !usesAutomaticRecording(cfg) || strings.TrimSpace(cfg.CanonicalSessionID) != "" {
+		return cfg, nil
+	}
+	generator := cfg.CanonicalSessionIDGenerator
+	if generator == nil {
+		generator = uuid.NewString
+	}
+	canonicalID := strings.TrimSpace(generator())
+	if canonicalID == "" {
+		return RunConfig{}, fmt.Errorf("canonical Factory Session ID generator returned an empty identity")
+	}
+	cfg.CanonicalSessionID = canonicalID
+	return cfg, nil
+}
+
+func usesAutomaticRecording(cfg RunConfig) bool {
+	return strings.TrimSpace(cfg.RecordPath) == "" &&
+		!cfg.DisableDefaultRecording &&
+		strings.TrimSpace(cfg.ReplayPath) == ""
+}
 
 // HostedInvocationOperation is the narrow capability retained by the CLI
 // after application opening. It is transport-local; the Factory Sessions root
